Refactor image parent getter
	* Extract image parent logic to it own module.
